requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
joblib
Pillow
plotly